graph/requirements.txt
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0